+++ resources/views/dashboard/search/users.blade.php
@@ -13,14 +13,19 @@
     <thead>
     <tr class="headings">
         <th>#  </th>
+        
         <th> @lang('site.user_id')</th>
+        
         <th> @lang('site.name')   </th>
-        <th> @lang('site.user_role')  </th>
+        
         <th> @lang('site.requests_count')  </th>
-        <th> @lang('site.rule') </th>        
+       
         <th> @lang('site.vip') </th>
+        
         <th> @lang('site.active') </th>
+        
         <th> @lang('site.registered_date') </th>
+        
         <th></th>
     </tr>
     </thead>
@@ -36,13 +41,8 @@
             <td>user#{{$item->id}}</td>
             
             <td>{{$item->name}}</td>
-            
-            <td><span class="label label-{{$item->user_role}}"> {{ __('site.'.$item->user_role) }}</span></td>
-            
+             
             <td> {{ $item->total_requests }} @lang('site.request') </td>
-
-            <td>{{ $item->rule ? $item->rule->name : '-' }}</td>
- 
 
             <td>
                 @if($item->vip ==1) <button class="btn btn-success btn-xs">
@@ -66,13 +66,16 @@
     
             <td>
                 <a class="whatsapp btn btn-success btn-xs" target="_blank" href="https://wa.me/966{{$item->mobile}}?text=
-                {{ \App\Models\Setting::getvalue('whatsapp_msg') }}">
-                <i class="fa fa-whatsapp"></i>
+                    {{ setting('whatsapp_msg') }}"> <i class="fa fa-whatsapp"></i>
                 </a>
 
-                <a href="{{ route('admin.user.edit',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @if(has_permission('users_edit'))
+                    <a href="{{ route('admin.user',$item->id) }}" class="btn btn-info btn-xs"> <i class="fa fa-edit"></i> </a>
+                @endif
 
-                <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @if(has_permission('users_delete'))
+                    <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs"><i class="fa fa-trash"></i> </a>
+                @endif
             </td>
         </tr>
            
@@ -89,24 +92,8 @@
 
 
 @section('scripts')
-    @include('dashboard.layouts.message') 
+    @include('dashboard.layouts.message_growl') 
 
     @include('dashboard.ajax.delete',['target'=>'user']) 
-
-    {{-- @include('dashboard.ajax.export_excel',['fileName'=>'users'])  --}}
-     
-    {{-- <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/1.3.2/jspdf.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf-autotable/2.3.2/jspdf.plugin.autotable.js"></script>
-    <script>
-        var doc = new jsPDF('p', 'pt');
-        var elem = document.getElementById("myTbl");
-        var res = doc.autoTableHtmlToJson(elem);
-
-        doc.autoTable(res.columns, res.data, {styles: {font: "Amiri"}});
-
-        // doc.autoTable(res.columns, res.data);
-        doc.save("table.pdf");
-    </script> --}}
-
-   
+ 
 @endsection
